# src/main/resources/swagger/api.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.0.1-SNAPSHOT",
    "title": "atlas",
    "x-appId": "1000001128",
    "x-serviceName": "atlas"
  },
  "host": "localhost:8081",
  "basePath": "/",
...
  "paths": {
...
    "/web/zones": {
      "post": {
        "tags": [
          "zone-controller"
        ],
        "summary": "createZone",
        "operationId": "createZoneUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "zoneDto",
            "description": "zoneDto",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ZoneDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«ZoneDto»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "zone-controller"
        ],
        "summary": "updateZone",
        "operationId": "updateZoneUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "zoneDto",
            "description": "zoneDto",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ZoneDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«ZoneDto»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/web/zones/all": {
      "get": {
        "tags": [
          "zone-controller"
        ],
        "summary": "getAllZones",
        "operationId": "getAllZonesUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«List«ZoneDto»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/web/zones/condition": {
      "get": {
        "tags": [
          "zone-controller"
        ],
        "summary": "searchZonesByEnvAndName",
        "operationId": "searchZonesByEnvAndNameUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "page",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "size",
            "in": "query",
            "description": "size",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "envName",
            "in": "query",
            "description": "envName",
            "required": true,
            "type": "string"
          },
          {
            "name": "zoneName",
            "in": "query",
            "description": "zoneName",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«PageDto»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/web/zones/getByEnv": {
      "get": {
        "tags": [
          "zone-controller"
        ],
        "summary": "getByEnv",
        "operationId": "getByEnvUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "env",
            "in": "query",
            "description": "env",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«List«ZoneDto»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/web/zones/page": {
      "get": {
        "tags": [
          "zone-controller"
        ],
        "summary": "getZonesInPage",
        "operationId": "getZonesInPageUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "page",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "size",
            "in": "query",
            "description": "size",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«PageDto»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/web/zones/{zoneId}": {
      "get": {
        "tags": [
          "zone-controller"
        ],
        "summary": "getZoneById",
        "operationId": "getZoneByIdUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "zoneId",
            "in": "path",
            "description": "zoneId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«ZoneDto»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "zone-controller"
        ],
        "summary": "deleteZone",
        "operationId": "deleteZoneUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "zoneId",
            "in": "path",
            "description": "zoneId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«string»"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    }
  },
  "definitions": {
...
    "PageDto": {
      "type": "object",
      "properties": {
        "content": {
          "type": "array",
          "items": {
            "type": "object"
          }
        },
        "first": {
          "type": "boolean"
        },
        "last": {
          "type": "boolean"
        },
        "number": {
          "type": "integer",
          "format": "int32"
        },
        "numberOfElements": {
          "type": "integer",
          "format": "int32"
        },
        "size": {
          "type": "integer",
          "format": "int32"
        },
        "totalElements": {
          "type": "integer",
          "format": "int64"
        },
        "totalPages": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
    "Response«List«ZoneDto»»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "detail": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ZoneDto"
          }
        },
        "message": {
          "type": "string"
        }
      }
    },
...
    "Response«PageDto»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "detail": {
          "$ref": "#/definitions/PageDto"
        },
        "message": {
          "type": "string"
        }
      }
    },
...
    "Response«ZoneDto»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "detail": {
          "$ref": "#/definitions/ZoneDto"
        },
        "message": {
          "type": "string"
        }
      }
    },
...
    "Response«string»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "detail": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
...
    "ZoneDto": {
      "type": "object",
      "properties": {
        "active": {
          "type": "boolean"
        },
        "description": {
          "type": "string"
        },
        "envId": {
          "type": "integer",
          "format": "int64"
        },
        "envName": {
          "type": "string"
        },
        "extensions": {
          "type": "object"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "insertBy": {
          "type": "string"
        },
        "insertTime": {
          "type": "string",
          "format": "date-time"
        },
        "k8s": {
          "type": "string"
        },
        "k8sVersion": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "updateBy": {
          "type": "string"
        },
        "updateTime": {
          "type": "string",
          "format": "date-time"
        }
      }
    }
  }
}